Refactored Protocol + Better `ServerPacket` Support

+++ src/main/java/me/outspending/entity/BlockEntity.java
@@ -2,4 +2,4 @@ package me.outspending.entity;
 
 import net.kyori.adventure.nbt.CompoundBinaryTag;
 
-public record BlockEntity(int packedXZ, short y, int type, CompoundBinaryTag data) {}
+public record BlockEntity(byte packedXZ, short y, int type, CompoundBinaryTag data) {}
